requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.6
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
PyJWT==2.8.0
python-barcode==0.15.1
pytz==2023.3.post1
pyzbar==0.1.9
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
